add ffmpeg_info, simple-ffmpeg-decoder, yuv_player

+++ libffmpeg/simple_ffmpeg_player.cpp
@@ -1,25 +1,24 @@
-/*
- * simple_ffmpeg_player.cpp --
- *
- * Copyright (C) 2016 liyunteng
- * Auther: liyunteng <li_yunteng@163.com>
- * License: GPL
- * Update time:  2016/11/29 23:46:27
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms and conditions of the GNU General Public License,
- * version 2, as published by the Free Software Foundation.
- *
- * This program is distributed in the hope it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin St - Fifth Floor, Boston, MA 02110-1301 USA.
- *
- */
+// -*- compile-command: "clang++ -Wall -o simple_ffmpeg_player simple_ffmpeg_player.cpp -g -lavformat -lavcodec -lswscale -lavutil -lSDL" -*-
+// simple_ffmpeg_player.cpp -- ffmepg player
+
+// Copyright (C) 2016 liyunteng
+// Auther: liyunteng <li_yunteng@163.com>
+// License: GPL
+// Update time:  2016/12/04 04:04:57
+
+// This program is free software; you can redistribute it and/or modify
+// it under the terms and conditions of the GNU General Public License
+// version 2,as published by the Free Software Foundation.
+
+// This program is distributed in the hope it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+// or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+// for more details.
+
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation,
+// Inc.,51 Franklin St - Fifth Floor, Boston,MA 02110-1301 USA.
+
 #include <stdio.h>
 
 #ifdef __cplusplus
@@ -71,7 +70,7 @@ int main(void)
         return -1;
     }
     videoIndex = -1;
-    for (i = 0; i < pFormatCtx->nb_streams; i++) {
+    for (i = 0; i < (int)pFormatCtx->nb_streams; i++) {
         if (pFormatCtx->streams[i]->codec->codec_type == AVMEDIA_TYPE_VIDEO) {
             videoIndex = i;
             break;
